Move subscription methd from Market to Terminal
Terminal class can fill this concern because DepthConsumeKey can filter
depths by any kind: exchange, symbol, or market.
